+++ clevertap-core/src/main/java/com/clevertap/android/sdk/Utils.java
@@ -613,9 +613,10 @@ public final class Utils {
         boolean exoPlayerPresent = false;
         Class className = null;
         try {
-            className = Class.forName("com.google.android.exoplayer2.SimpleExoPlayer");
+            className = Class.forName("com.google.android.exoplayer2.ExoPlayer");
             className = Class.forName("com.google.android.exoplayer2.source.hls.HlsMediaSource");
-            className = Class.forName("com.google.android.exoplayer2.ui.PlayerView");
+            className = Class.forName("com.google.android.exoplayer2.ui.StyledPlayerView");
+
             Logger.d("ExoPlayer is present");
             exoPlayerPresent = true;
         } catch (Throwable t) {