+++ clevertap-core/src/main/java/com/clevertap/android/sdk/inbox/CTInboxBaseMessageViewHolder.java
@@ -1,13 +1,10 @@
 package com.clevertap.android.sdk.inbox;
 
-import static com.google.android.exoplayer2.ui.PlayerView.SHOW_BUFFERING_NEVER;
-
 import android.annotation.SuppressLint;
 import android.content.Context;
 import android.content.res.Configuration;
 import android.content.res.Resources;
 import android.graphics.Color;
-import android.net.Uri;
 import android.util.DisplayMetrics;
 import android.util.TypedValue;
 import android.view.Gravity;
@@ -18,18 +15,22 @@ import android.widget.FrameLayout;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.RelativeLayout;
+
 import androidx.annotation.NonNull;
 import androidx.annotation.RestrictTo;
 import androidx.annotation.RestrictTo.Scope;
 import androidx.core.content.res.ResourcesCompat;
 import androidx.recyclerview.widget.RecyclerView;
+
 import com.clevertap.android.sdk.R;
-import com.google.android.exoplayer2.SimpleExoPlayer;
+import com.google.android.exoplayer2.ExoPlayer;
+import com.google.android.exoplayer2.MediaItem;
 import com.google.android.exoplayer2.source.hls.HlsMediaSource;
-import com.google.android.exoplayer2.ui.PlayerView;
+import com.google.android.exoplayer2.ui.StyledPlayerView;
 import com.google.android.exoplayer2.upstream.DataSource;
 import com.google.android.exoplayer2.upstream.DefaultBandwidthMeter;
-import com.google.android.exoplayer2.upstream.DefaultDataSourceFactory;
+import com.google.android.exoplayer2.upstream.DefaultDataSource;
+import com.google.android.exoplayer2.upstream.DefaultHttpDataSource;
 import com.google.android.exoplayer2.util.Util;
 import java.lang.ref.WeakReference;
 import java.text.SimpleDateFormat;
@@ -66,7 +67,7 @@ public class CTInboxBaseMessageViewHolder extends RecyclerView.ViewHolder {
         super(itemView);
     }
 
-    public boolean addMediaPlayer(PlayerView videoSurfaceView) {
+    public boolean addMediaPlayer(StyledPlayerView videoSurfaceView) {
         if (!requiresMediaPlayer) {
             return false;
         }
@@ -105,7 +106,7 @@ public class CTInboxBaseMessageViewHolder extends RecyclerView.ViewHolder {
             progressBarFrameLayout.setVisibility(View.VISIBLE);
         }
 
-        final SimpleExoPlayer player = (SimpleExoPlayer) videoSurfaceView.getPlayer();
+        final ExoPlayer player =(ExoPlayer) videoSurfaceView.getPlayer();
         float currentVolume = 0;
         if (player != null) {
             currentVolume = player.getVolume();
@@ -129,41 +130,44 @@ public class CTInboxBaseMessageViewHolder extends RecyclerView.ViewHolder {
             layoutParams.setMargins(0, iconTop, iconRight, 0);
             layoutParams.gravity = Gravity.END;
             muteIcon.setLayoutParams(layoutParams);
-            muteIcon.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    float currentVolume = 0;
+            muteIcon.setOnClickListener(v -> {
+                float currentVolume1 = 0;
+                if (player != null) {
+                    currentVolume1 = player.getVolume();
+                }
+                if (currentVolume1 > 0) {
+                    player.setVolume(0f);
+                    muteIcon.setImageDrawable(
+                            ResourcesCompat.getDrawable(context.getResources(), R.drawable.ct_volume_off, null));
+                } else if (currentVolume1 == 0) {
                     if (player != null) {
-                        currentVolume = player.getVolume();
+                        player.setVolume(1);
                     }
-                    if (currentVolume > 0) {
-                        player.setVolume(0f);
-                        muteIcon.setImageDrawable(
-                                ResourcesCompat.getDrawable(context.getResources(), R.drawable.ct_volume_off, null));
-                    } else if (currentVolume == 0) {
-                        if (player != null) {
-                            player.setVolume(1);
-                        }
-                        muteIcon.setImageDrawable(
-                                ResourcesCompat.getDrawable(context.getResources(), R.drawable.ct_volume_on, null));
-                    }
+                    muteIcon.setImageDrawable(
+                            ResourcesCompat.getDrawable(context.getResources(), R.drawable.ct_volume_on, null));
                 }
             });
             frameLayout.addView(muteIcon);
         }
 
         videoSurfaceView.requestFocus();
-        videoSurfaceView.setShowBuffering(SHOW_BUFFERING_NEVER);
+        videoSurfaceView.setShowBuffering(StyledPlayerView.SHOW_BUFFERING_NEVER);
         DefaultBandwidthMeter defaultBandwidthMeter = new DefaultBandwidthMeter.Builder(context).build();
-        DataSource.Factory dataSourceFactory = new DefaultDataSourceFactory(context,
-                Util.getUserAgent(context, context.getPackageName()), defaultBandwidthMeter);
+
+        Context ctx = this.context;
+        String userAgent = Util.getUserAgent(ctx,ctx.getPackageName());
         String uriString = firstContentItem.getMedia();
+        MediaItem mediaItem = MediaItem.fromUri(uriString);
+        DefaultHttpDataSource.Factory  dsf = new DefaultHttpDataSource.Factory().setUserAgent(userAgent).setTransferListener(defaultBandwidthMeter);
+        DataSource.Factory dataSourceFactory = new DefaultDataSource.Factory(ctx,dsf);
+
+
         if (uriString != null) {
-            HlsMediaSource hlsMediaSource = new HlsMediaSource.Factory(dataSourceFactory)
-                    .createMediaSource(Uri.parse(uriString));
+            HlsMediaSource hlsMediaSource = new HlsMediaSource.Factory(dataSourceFactory).createMediaSource(mediaItem);
             // Prepare the player with the source.
             if (player != null) {
-                player.prepare(hlsMediaSource);
+                player.setMediaSource(hlsMediaSource);
+                player.prepare();
                 if (firstContentItem.mediaIsAudio()) {
                     videoSurfaceView.showController();//show controller for audio as it is not autoplay
                     player.setPlayWhenReady(false);